requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.56.1
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
